--- server/config/custom.d.ts
@@ -1,5 +1,0 @@
-declare namespace Express {
-  export interface Request {
-     decoded?:string
-  }
-}
